nav mesh script is working.
this is a working version of navmesh task for whom want to use navmesh script (you can call the function NavMeshProvider(),see the comment in the script). it does not call locomotinon function cuz locomotion team needs to fix issues about movement direction. 
this version include two baked navmesh map. one is for humanoid, one is for gaint.